+++ src/packages/responsive-navigation/MobileNavigation.js
@@ -1,0 +1,32 @@
+import { useEffect, useState } from 'react';
+import PropTypes from 'prop-types';
+
+const MobileNavigation = ({ activePath, children, history }) => {
+  const [isNavigationVisible, setIsNavigationVisible] = useState(false);
+
+  function hideNavigation() {
+    setIsNavigationVisible(false);
+  }
+
+  function showNavigation() {
+    setIsNavigationVisible(true);
+  }
+
+  useEffect(() => history.listen(hideNavigation), []);
+
+  return children({
+    activePath,
+    hideNavigation,
+    showNavigation,
+    isMobile: true,
+    isNavigationVisible,
+  });
+};
+
+MobileNavigation.propTypes = {
+  activePath: PropTypes.string.isRequired,
+  children: PropTypes.func.isRequired,
+  history: PropTypes.object.isRequired,
+};
+
+export default MobileNavigation;
